- commit message : """ rename : Entities -> Entity 폴더명 변경 issue: [rename] Modify from 'Entities' Folder to 'Entity' Folder #8
---

refactor: id 컬럼 bigint -> string으로 변경
issue: [refactor] from ({ type : 'bigint'}) to ({ type : 'string' }) #12

---

refactor: from 'like.toilet_id' to 'like.toilet'

추가적인 설명 :
중간에 commit을 못해서
수정된 부분이 마킹되지 않았습니다.
(17line)이 수정된 부분입니다.

"""
